backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.4
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
requests==2.28.1
beautifulsoup4==4.11.1
pytest==7.1.2
pytest-flask==1.2.0
pytest-cov==4.1.0
psycopg2-binary>=2.9.9,<3.0
gunicorn==21.2.0
SQLAlchemy==1.4.52
python-dotenv==1.0.0
selenium==4.15.2
argon2-cffi==23.2.0
# Flask 2.2 is not fully compatible with Werkzeug 3.x; pin to <3 to avoid runtime errors
Werkzeug<3.0